refactor(config): move pixel classifiers to channel config and add region filter
