+++ src/main/java/qupath/ext/braian/config/ProjectsConfig.java
@@ -98,11 +98,8 @@ public class ProjectsConfig {
 
     private String classForDetections = null;
     private String atlasName = "allen_mouse_10um_java";
-    private boolean enableCellDetection = true;
-    private boolean enablePixelClassification = false;
     private DetectionsCheckConfig detectionsCheck = new DetectionsCheckConfig();
     private List<ChannelDetectionsConfig> channelDetections = List.of();
-    private List<PixelClassifierConfig> pixelClassifiers = List.of();
 
     public String getClassForDetections() {
         return classForDetections;
@@ -143,22 +140,6 @@ public class ProjectsConfig {
         this.atlasName = atlasName;
     }
 
-    public boolean isEnableCellDetection() {
-        return enableCellDetection;
-    }
-
-    public void setEnableCellDetection(boolean enableCellDetection) {
-        this.enableCellDetection = enableCellDetection;
-    }
-
-    public boolean isEnablePixelClassification() {
-        return enablePixelClassification;
-    }
-
-    public void setEnablePixelClassification(boolean enablePixelClassification) {
-        this.enablePixelClassification = enablePixelClassification;
-    }
-
     public DetectionsCheckConfig getDetectionsCheck() {
         return detectionsCheck;
     }
@@ -193,11 +174,4 @@ public class ProjectsConfig {
         this.channelDetections = channelDetections;
     }
 
-    public List<PixelClassifierConfig> getPixelClassifiers() {
-        return pixelClassifiers;
-    }
-
-    public void setPixelClassifiers(List<PixelClassifierConfig> pixelClassifiers) {
-        this.pixelClassifiers = pixelClassifiers;
-    }
 }
